+++ CHANGELOG.md
@@ -3,3 +3,4 @@
 ## 3.2.5.5
 
 -   Fix issue with skins not downloading causes application hang
+-   Fix issue with some Forge loader libraries not installing for servers